+++ apps/judicial-system/web/pages/krafa/domkrofur-og-lagaakvaedi/[id].tsx
@@ -1,0 +1,3 @@
+import { StepThree } from '@island.is/judicial-system-web/src/routes/Prosecutor/CreateDetentionRequest/StepThree/StepThree'
+
+export default StepThree
