+++ apps/judicial-system/web/pages/krafa/fyrirtaka/[id].tsx
@@ -1,0 +1,3 @@
+import { StepTwo } from '@island.is/judicial-system-web/src/routes/Prosecutor/CreateDetentionRequest/StepTwo/StepTwo'
+
+export default StepTwo
