+++ apps/judicial-system/web/pages/krafa/sakborningur/[id].tsx
@@ -1,0 +1,3 @@
+import StepOne from '@island.is/judicial-system-web/src/routes/Prosecutor/CreateDetentionRequest/StepOne/StepOne'
+
+export default StepOne
