+++ apps/judicial-system/web/pages/krofur.tsx
@@ -1,0 +1,3 @@
+import DetentionRequests from '@island.is/judicial-system-web/src/routes/Shared/DetentionRequests/DetentionRequests'
+
+export default DetentionRequests
